Add initial support for scheduling and applying the configuration changes
Also, add some plain events objects that will be used by higher layer
code instead of the default web3 logs objects.

+++ solidity/contracts/Gatekeeper.sol
@@ -96,7 +96,7 @@ contract Gatekeeper is DelayedOps, PermissionsLevel {
     uint constant maxDelay = 365 days;
     uint constant maxFreeze = 365 days;
 
-    function initialConfig(Vault vaultParam, bytes32[] memory initialParticipants, uint[] memory initialDelays) public {
+    function initialConfig(Vault vaultParam, bytes32[] memory initialParticipants, uint256[] memory initialDelays) public {
         require(operator == address(0), "already initialized");
 
         require(initialParticipants.length <= maxParticipants, "too many participants");
